requirements.txt
jieba3k
flask
pymongo
lxml
